# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * EliteCare Clinic API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type DoctorAvailabilityStatus = typeof DoctorAvailabilityStatus[keyof typeof DoctorAvailabilityStatus];


export const DoctorAvailabilityStatus = {
  available: 'available',
  next_slot: 'next_slot',
  unavailable: 'unavailable',
} as const;

export interface Doctor {
  id: number;
  name: string;
  specialty: string;
  qualification: string;
  experience: string;
  rating: number;
  patientsCount?: number;
  /** @nullable */
  image: string | null;
  availabilityStatus: DoctorAvailabilityStatus;
  /** @nullable */
  nextAvailableTime?: string | null;
  /** @nullable */
  bio?: string | null;
}

export type AppointmentConsultationType = typeof AppointmentConsultationType[keyof typeof AppointmentConsultationType];


export const AppointmentConsultationType = {
  in_person: 'in_person',
  online: 'online',
} as const;

export type AppointmentStatus = typeof AppointmentStatus[keyof typeof AppointmentStatus];


export const AppointmentStatus = {
  pending: 'pending',
  confirmed: 'confirmed',
  completed: 'completed',
  cancelled: 'cancelled',
} as const;

export interface Appointment {
  id: number;
  patientName: string;
  patientEmail?: string;
  patientPhone?: string;
  doctorId: number;
  doctorName?: string;
  date: string;
  timeSlot: string;
  reason: string;
  consultationType?: AppointmentConsultationType;
  status: AppointmentStatus;
  createdAt: string;
}

export type AppointmentInputConsultationType = typeof AppointmentInputConsultationType[keyof typeof AppointmentInputConsultationType];


export const AppointmentInputConsultationType = {
  in_person: 'in_person',
  online: 'online',
} as const;

export interface AppointmentInput {
  patientName: string;
  patientEmail: string;
  patientPhone: string;
  doctorId: number;
  date: string;
  timeSlot: string;
  reason: string;
  consultationType: AppointmentInputConsultationType;
}

export interface ClinicStats {
  happyPatients: number;
  yearsExperience: number;
  successRate: number;
  emergencySupport: string;
  doctorsCount: number;
}

export interface Testimonial {
  id: number;
  patientName: string;
  rating: number;
  review: string;
  treatment: string;
  date: string;
  /** @nullable */
  avatarUrl?: string | null;
}

export interface ChatMessageInput {
  message: string;
  /** @nullable */
  sessionId?: string | null;
}

export interface ChatResponse {
  reply: string;
  sessionId: string;
  suggestedActions?: string[];
}

export type GetAvailableSlotsParams = {
doctorId?: number;
date?: string;
};