# lib/api-zod/src/generated/types/appointment.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * EliteCare Clinic API
 * OpenAPI spec version: 0.1.0
 */
import type { AppointmentConsultationType } from './appointmentConsultationType';
import type { AppointmentStatus } from './appointmentStatus';

export interface Appointment {
  id: number;
  patientName: string;
  patientEmail?: string;
  patientPhone?: string;
  doctorId: number;
  doctorName?: string;
  date: string;
  timeSlot: string;
  reason: string;
  consultationType?: AppointmentConsultationType;
  status: AppointmentStatus;
  createdAt: string;
}