# lib/api-zod/src/generated/types/appointmentInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * EliteCare Clinic API
 * OpenAPI spec version: 0.1.0
 */
import type { AppointmentInputConsultationType } from './appointmentInputConsultationType';

export interface AppointmentInput {
  patientName: string;
  patientEmail: string;
  patientPhone: string;
  doctorId: number;
  date: string;
  timeSlot: string;
  reason: string;
  consultationType: AppointmentInputConsultationType;
}