# lib/api-zod/src/generated/types/doctor.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * EliteCare Clinic API
 * OpenAPI spec version: 0.1.0
 */
import type { DoctorAvailabilityStatus } from './doctorAvailabilityStatus';

export interface Doctor {
  id: number;
  name: string;
  specialty: string;
  qualification: string;
  experience: string;
  rating: number;
  patientsCount?: number;
  /** @nullable */
  image: string | null;
  availabilityStatus: DoctorAvailabilityStatus;
  /** @nullable */
  nextAvailableTime?: string | null;
  /** @nullable */
  bio?: string | null;
}