# lib/api-zod/src/generated/types/testimonial.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * EliteCare Clinic API
 * OpenAPI spec version: 0.1.0
 */

export interface Testimonial {
  id: number;
  patientName: string;
  rating: number;
  review: string;
  treatment: string;
  date: string;
  /** @nullable */
  avatarUrl?: string | null;
}